Bump xmtp-ffi to 0.1.8, remove redundant as i64 cast, flatten nested if-let with let-chain, and remove unnecessary Default::default() spread
Increment version to 0.1.8. Remove `as i64` cast from `now_ns()` call in client.rs (function already returns i64). Refactor nested `if let Ok(ids)` + `if !ids.is_empty()` in conversation.rs into single let-chain pattern (`if let Ok(ids) = ids && !ids.is_empty()`). Remove `..Default::default()` spread from GroupMetadataOptions construction in conversations.rs (

+++ xmtp-ffi/src/client.rs
@@ -318,7 +318,7 @@ pub unsafe extern "C" fn xmtp_client_set_consent_states(
             return Err("null pointer or invalid count".into());
         }
 
-        let now_ns = xmtp_common::time::now_ns() as i64;
+        let now_ns = xmtp_common::time::now_ns();
         let mut records = Vec::with_capacity(count as usize);
         for i in 0..count as usize {
             let entity = unsafe { c_str_to_string(*entities.add(i))? };
